Move DatabaseContents to separate file
